+++ src/main/java/com/ll/goohaeyou/domain/calculate/JobListener.java
@@ -1,0 +1,23 @@
+package com.ll.goohaeyou.domain.calculate;
+
+import org.springframework.batch.core.JobExecution;
+import org.springframework.batch.core.JobExecutionListener;
+import org.springframework.stereotype.Component;
+
+@Component
+public class JobListener implements JobExecutionListener {
+    private long startTime;
+    private long endTime;
+
+    @Override
+    public void beforeJob(JobExecution jobExecution) {
+        startTime = System.currentTimeMillis();
+    }
+
+    @Override
+    public void afterJob(JobExecution jobExecution) {
+        endTime = System.currentTimeMillis();
+        long executionTime = endTime - startTime;
+        jobExecution.getExecutionContext().putLong("executionTime", executionTime);
+    }
+}
