--- src/main/java/com/ll/goohaeyou/domain/calculate/calculate/controller/BatchController.java
@@ -1,36 +1,0 @@
-package com.ll.goohaeyou.domain.calculate.calculate.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.batch.core.Job;
-import org.springframework.batch.core.JobExecution;
-import org.springframework.batch.core.JobParameters;
-import org.springframework.batch.core.JobParametersBuilder;
-import org.springframework.batch.core.launch.JobLauncher;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequiredArgsConstructor
-public class BatchController {
-    private final Job exceeded3DaysJob;
-    private final JobLauncher jobLauncher;
-
-    @PostMapping("/batch")
-    // 개발과정에서 확인을 위해 작성
-    public ResponseEntity<String> runBatch() {
-        try {
-            JobParameters jobParameters = new JobParametersBuilder()
-                    .addLong("time", System.currentTimeMillis())
-                    .toJobParameters();
-
-            JobExecution jobExecution = jobLauncher.run(exceeded3DaysJob, jobParameters);
-
-            return ResponseEntity.ok("Batch job started with ID: " + jobExecution.getId());
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("Failed to start batch job: " + e.getMessage());
-        }
-    }
-}
