--- src/main/java/com/ll/goohaeyou/domain/member/review/dto/ApplicantReviewDto.java
@@ -1,14 +1,0 @@
-package com.ll.goohaeyou.domain.member.review.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class ApplicantReviewDto {
-    private Long id;
-    private String body;
-    private double score;
-    private Long jobPostingId;
-    private Long applicantId;
-}
