+++ src/main/java/com/ll/goohaeyou/domain/review/dto/ApplicantReviewDto.java
@@ -1,0 +1,14 @@
+package com.ll.goohaeyou.domain.review.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class ApplicantReviewDto {
+    private Long id;
+    private String body;
+    private double score;
+    private Long jobPostingId;
+    private Long applicantId;
+}
