+++ src/main/java/com/ll/goohaeyou/global/event/notification/ApplicationCreateAndChangedEvent.java
@@ -1,6 +1,6 @@
 package com.ll.goohaeyou.global.event.notification;
 
-import com.ll.goohaeyou.domain.application.application.entity.Application;
+import com.ll.goohaeyou.domain.application.entity.Application;
 import com.ll.goohaeyou.domain.jobPost.jobPost.entity.JobPostDetail;
 import com.ll.goohaeyou.domain.notification.entity.type.CauseTypeCode;
 import lombok.Getter;
